+++ lib/nextgen_button.dart
@@ -1,3 +1,1 @@
-library nextgen_button;
-
 export 'src/nextgen_button.dart';