+++ test/nextgen_button_test.dart
@@ -1,5 +1,1 @@
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:nextgen_button/nextgen_button.dart';
-
 void main() {}